+++ tests/util/mod.rs
@@ -48,17 +48,11 @@ pub fn test_openapi_response(server : &TestServer, path : &str, output_file : &s
 {
 	info!("GET {}", path);
 	let res = server.client().get(path).perform().unwrap().read_body().unwrap();
-	let body = serde_json::to_string_pretty(&serde_json::from_slice::<serde_json::Value>(res.as_ref()).unwrap()).unwrap();
-	match File::open(output_file) {
-		Ok(mut file) => {
-			let mut expected = String::new();
-			file.read_to_string(&mut expected).unwrap();
-			eprintln!("{}", body);
-			assert_eq!(body, expected);
-		},
-		Err(_) => {
-			let mut file = File::create(output_file).unwrap();
-			file.write_all(body.as_bytes()).unwrap();
-		}
-	};
+	let body: serde_json::Value = serde_json::from_slice(&res).unwrap();
+
+	let mut file = File::open(output_file).unwrap();
+	let expected: serde_json::Value = serde_json::from_reader(&mut file).unwrap();
+
+	//eprintln!("{}", body);
+	assert_eq!(body, expected);
 }